[monitoring/uss_qualifier] move rid components to package rid
