+++ monitoring/uss_qualifier/rid/main.py
@@ -7,7 +7,7 @@ import argparse
 from urllib.parse import urlparse
 from monitoring.monitorlib.typing import ImplicitDict
 from monitoring.uss_qualifier.utils import RIDQualifierTestConfiguration
-import monitoring.uss_qualifier.test_executor as test_executor
+import monitoring.uss_qualifier.rid.test_executor as test_executor
 
 def is_url(url_string):
     try:
