+++ monitoring/uss_qualifier/rid/simulator/flight_state_from_kml_test.py
@@ -10,19 +10,19 @@ PACKAGE='monitoring.uss_qualifier.rid.simulator'
 
 def test_get_interpolated_value(mocker):
     mocker.patch(
-        f'{PACKAGE}.flight_data_from_kml.get_polygons_distances_from_point',
+        f'{PACKAGE}.flight_state_from_kml.get_polygons_distances_from_point',
         return_value=[1.1, 2.0, 3.3, 0.0])
     result_alt = frk.get_interpolated_value('point', 'polygons', [10,20,30, 40])
     assert result_alt == 40
 
     mocker.patch(
-        f'{PACKAGE}.flight_data_from_kml.get_polygons_distances_from_point',
+        f'{PACKAGE}.flight_state_from_kml.get_polygons_distances_from_point',
         return_value=[10, 50, 100])
     result_alt = frk.get_interpolated_value('point', 'polygons', [10,20,30])
     assert abs(result_alt - 13.1) < 0.1
 
     mocker.patch(
-        f'{PACKAGE}.flight_data_from_kml.get_polygons_distances_from_point',
+        f'{PACKAGE}.flight_state_from_kml.get_polygons_distances_from_point',
         return_value=[10, 50, 100])
     result_alt = frk.get_interpolated_value('point', 'polygons', [140,125,116])
     assert abs(result_alt - 135.84) < 0.1
